+++ app/helpers/evaluation_helper.rb
@@ -1,12 +1,25 @@
 # coding: utf-8
 
 module EvaluationHelper
+  def evaluation_table_headers
+    [
+      {name: '実行開始日時', width: 25},
+      {name: 'モデル', width: 20},
+      {name: '状態', width: 15},
+      {name: '適合率', width: 10},
+      {name: '再現率', width: 10},
+      {name: 'F値', width: 10},
+    ]
+  end
+
   def progress(evaluation)
     case evaluation.state
     when 'completed'
       '完了'
     when 'error'
       'エラー'
+    when 'waiting'
+      '実行待ち'
     else
       if evaluation.data.empty?
         '0%完了'
@@ -34,6 +47,17 @@ module EvaluationHelper
       'Top20' => 'remote',
       'ファイル' => 'file',
       '直接入力' => 'text',
+      'ランダム' => 'random',
     }
+  end
+
+  def download_button(evaluation)
+    return if %w[text file].include?(evaluation.data_source)
+    return unless evaluation.state == 'completed'
+
+    id = evaluation.evaluation_id
+    content_tag(:button, id: id, class: 'btn btn-success btn-download') do
+      content_tag(:span, nil, class: 'glyphicon glyphicon-download-alt')
+    end
   end
 end