+++ app/models/prediction.rb
@@ -1,10 +1,26 @@
 class Prediction < ApplicationRecord
+  include ModelUtil
+
   validates :model, :test_data, :state,
             presence: {message: 'absent'}
   validates :state,
-            inclusion: {in: %w[processing completed error], message: 'invalid'}
+            inclusion: {in: STATE_LIST, message: 'invalid'},
+            allow_nil: true
 
   has_many :results, as: :predictable, dependent: :destroy, inverse_of: :predictable
+
+  after_initialize if: :new_record? do |prediction|
+    prediction.state = DEFAULT_STATE
+  end
+
+  def set_analysis!
+    data_dir = Rails.root.join('tmp', 'files', id.to_s)
+    analysis_id = read_analysis_id(File.join(data_dir, 'metadata.yml'))
+    analysis = Analysis.find_by(analysis_id: analysis_id)
+    raise StandardError if analysis.nil?
+
+    update!(analysis: analysis)
+  end
 
   def import_results(result_file)
     race_result = YAML.load_file(result_file)