+++ config/routes.rb
@@ -7,5 +7,8 @@ Rails.application.routes.draw do
 
   get '/evaluations' => 'evaluations#manage'
   post '/evaluations' => 'evaluations#execute'
-  resources :evaluations, only: %i[show]
+  resources :evaluations, only: %i[] do
+    get 'download' => 'evaluations#download', param: :evaluation_id
+  end
+  resources :evaluations, only: %i[show], param: :evaluation_id
 end