+++ db/migrate/20200101134947_add_analysis_id_to_evaluations.rb
@@ -1,0 +1,5 @@
+class AddAnalysisIdToEvaluations < ActiveRecord::Migration[5.0]
+  def change
+    add_reference :evaluations, :analysis, first: true
+  end
+end