+++ scripts/analyze_with_num_entry.py
@@ -1,5 +1,5 @@
-from dtreeviz.trees import dtreeviz
 from sklearn.ensemble import RandomForestClassifier
+import analysis_util as util
 import mysql.connector as mysql
 import numpy as np
 import os
@@ -15,7 +15,7 @@ ntree = int(args[3])
 nentry = int(args[4])
 
 workdir = os.path.dirname(os.path.abspath(args[0]))
-outputdir = workdir + '/../tmp/files/' + analysis_id
+outputdir = workdir + '/../tmp/files/analyses/' + analysis_id
 config = yaml.load(open(workdir + '/../config/settings.yml', 'r+'))
 
 def create_race_feature(group):
@@ -41,6 +41,8 @@ def create_race_feature(group):
 
   for i in range(nentry):
     feature = feature.drop('won_' + str(i), axis=1)
+
+  feature['race_id'] = group.iloc[0]['race_id']
 
   return feature
 
@@ -73,12 +75,24 @@ mapping = yaml.load(open(workdir + '/mapping.yml', 'r+'))
 for name in mapping:
   feature[name] = feature[name].map(mapping[name]).astype(int)
 
-feature.to_csv(outputdir + '/feature.csv', index=False)
+columns = feature.columns.to_list()
+columns.remove('race_id')
+columns.remove('number')
+columns.insert(0, 'number')
+columns.insert(0, 'race_id')
+feature[columns].sort_values(['race_id', 'number']).to_csv(outputdir + '/feature.csv', index=False)
+
 training_data = feature.groupby('race_id').apply(create_race_feature)
-training_data.to_csv(outputdir + '/training_data.csv', index=False)
+training_data = training_data.dropna()
+columns = training_data.columns.to_list()
+columns.remove('race_id')
+columns.insert(0, 'race_id')
+training_data[columns].to_csv(outputdir + '/training_data.csv', index=False)
+
+training_data = training_data.drop('race_id', axis=1)
 
 classifier = RandomForestClassifier(n_estimators=ntree, random_state=0)
-classifier.fit(training_data.drop('won', axis=1), training_data['won'])
+classifier.fit(training_data.drop('won', axis=1), training_data['won'].astype('int'))
 
 file = open(outputdir + '/metadata.yml', 'w+')
 importance_values = classifier.feature_importances_.astype(type('float', (float,), {}))
@@ -95,14 +109,4 @@ metadata = {
 file.write(yaml.dump(metadata))
 
 pickle.dump(classifier, open(outputdir + '/model.rf', 'wb'))
-
-#for i, estimator in enumerate(classifier.estimators_):
-#  tree = dtreeviz(
-#    estimator,
-#    training_data.drop('won', axis=1),
-#    training_data['won'],
-#    target_name='Result',
-#    feature_names=training_data.drop('won', axis=1).columns,
-#    class_names=range(nentry),
-#  )
-#  tree.save(outputdir + '/tree_' + str(i) + '.svg')
+util.output_tree(classifier.estimators_, training_data, outputdir)