+++ spec/factories/analysis.rb
@@ -1,8 +1,9 @@
 FactoryBot.define do
   factory :analysis do
+    analysis_id { '0' * 32 }
     num_data { 10000 }
     num_tree { 100 }
     num_feature { 20 }
-    state { 'processing' }
+    state { 'waiting' }
   end
 end