+++ spec/factories/evaluation.rb
@@ -3,7 +3,8 @@ FactoryBot.define do
     evaluation_id { '0' * 32 }
     model { 'model.rf' }
     data_source { 'remote' }
-    state { 'processing' }
+    num_data { 20 }
+    state { 'waiting' }
     precision { 0.5 }
     recall { 0.5 }
     f_measure { 0.5 }